Split UI to several files
